+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/params/ProcessParams.java
@@ -70,7 +70,9 @@ public record ProcessParams(
                 geometryParams().isSaturatedDiskMode(),
                 geometryParams().referencePolynomialDirectory().orElse(null),
                 geometryParams().isSpectrumVFlip(),
-                geometryParams().ellipseFittingMode()
+                geometryParams().ellipseFittingMode(),
+                geometryParams().horizontalFlipCondition().orElse(null),
+                geometryParams().verticalFlipCondition().orElse(null)
             ),
             bandingCorrectionParams,
             requestedImages,